Update README, utils API, and Makefile
Rewrite README with clearer usage and build instructions and add art examples. Change get_llen_scale signature from float to size_t (header, utils.c) and fix scale calculation cast. Update main.c to parse custom line length as size_t (using abs on atoi), adjust error return codes, and clean up comments. Tidy Makefile (remove extra blank, simplify clean target). These changes improve API correctness, formatting, and documentation.

+++ include/utils.h
@@ -4,6 +4,6 @@
 #include <stddef.h>
 
 char *image_2_ascii(const unsigned char *image,const int width,const int height,const int num_channel);
-float get_llen_scale(size_t width, float llen);
+float get_llen_scale(size_t width, size_t llen);
 
 #endif
